# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Readora API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /books:
    get:
      operationId: listBooks
      tags: [books]
      summary: List user books
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: genreId
          in: query
          schema:
            type: integer
        - name: author
          in: query
          schema:
            type: string
        - name: cycleId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [not_started, reading, finished]
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [title, author, uploadedAt, lastReadAt, progress]
        - name: sortDir
          in: query
          schema:
            type: string
            enum: [asc, desc]
        - name: groupBy
          in: query
          schema:
            type: string
            enum: [author, genre, cycle, status, none]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Book"
  /books/upload:
    post:
      operationId: uploadBook
      tags: [books]
      summary: Upload a book file (FB2 or EPUB)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/BookUploadInput"
      responses:
        "201":
          description: Uploaded and parsed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
        "400":
          description: Invalid file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /books/{id}:
    get:
      operationId: getBook
      tags: [books]
      summary: Get book details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateBook
      tags: [books]
      summary: Update book metadata
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
    delete:
      operationId: deleteBook
      tags: [books]
      summary: Delete a book
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /books/delete-bulk:
    post:
      operationId: deleteBulkBooks
      tags: [books]
      summary: Delete multiple books
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkDeleteInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkDeleteResult"

  # READER
...
  /public/popular-books:
    get:
      operationId: getPopularBooks
      tags: [books]
      summary: Get popular books for home page (public metadata only)
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PublicBook"

  # ADMIN
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # AUTH SCHEMAS
...
    Genre:
      type: object
      required: [id, code, name, isActive]
      properties:
        id:
          type: integer
        code:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        isActive:
          type: boolean
...
    Book:
      type: object
      required: [id, title, format, status, uploadedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        author:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        coverUrl:
          type: ["string", "null"]
        format:
          type: string
          enum: [fb2, epub]
        language:
          type: ["string", "null"]
        publicationYear:
          type: ["integer", "null"]
        status:
          type: string
          enum: [active, blocked]
        cycleId:
          type: ["integer", "null"]
        cycleName:
          type: ["string", "null"]
        cycleNumber:
          type: ["number", "null"]
        fileSize:
          type: integer
        wordCount:
          type: ["integer", "null"]
        chapterCount:
          type: ["integer", "null"]
        genres:
          type: array
          items:
            $ref: "#/components/schemas/Genre"
        readingStatus:
          type: ["string", "null"]
          enum: [not_started, reading, finished, null]
        progressPercent:
          type: ["number", "null"]
        lastReadAt:
          type: ["string", "null"]
          format: date-time
        uploadedAt:
          type: string
          format: date-time
    BookUploadInput:
      type: object
      properties:
        cycleId:
          type: integer
        cycleName:
          type: string
        cycleNumber:
          type: number
    BookUpdate:
      type: object
      properties:
        title:
          type: string
        author:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        language:
          type: ["string", "null"]
        publicationYear:
          type: ["integer", "null"]
        cycleId:
          type: ["integer", "null"]
        cycleNumber:
          type: ["number", "null"]
        genreIds:
          type: array
          items:
            type: integer
    BulkDeleteInput:
      type: object
      required: [ids]
      properties:
        ids:
          type: array
          items:
            type: integer
    BulkDeleteResult:
      type: object
      required: [deleted]
      properties:
        deleted:
          type: integer
    PublicBook:
      type: object
      required: [id, title]
      properties:
        id:
          type: integer
        title:
          type: string
        author:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        coverUrl:
          type: ["string", "null"]
        openCount:
          type: integer

    # READER
...